+++ docs/source/changelog.rst
@@ -18,6 +18,7 @@ v26.1.12-dev
 - Added :class:`~discord.events.MessageUpdateEvent` event.
 - Added :class:`~discord.objects.Message` object.
 - Added :class:`~discord.objects.SharedClientTheme` object.
+- Added :class:`~discord.objects.BaseThemeType` object.
 
 
 v26.1.11